Move Cognito to Core
