+++ core/go.mod
@@ -3,9 +3,13 @@ module github.com/robinje/multi-user-dungeon/core
 go 1.22
 
 require (
+	github.com/aws/aws-sdk-go v1.54.15
 	github.com/google/uuid v1.6.0
 	go.etcd.io/bbolt v1.3.10
 	golang.org/x/crypto v0.24.0
 )
 
-require golang.org/x/sys v0.21.0 // indirect
+require (
+	github.com/jmespath/go-jmespath v0.4.0 // indirect
+	golang.org/x/sys v0.21.0 // indirect
+)
